extract-build-kv-cache-move: Move build_kv_cache to mem_cache.kv_cache_builder
Mechanical cut + paste for the ``extract-build-kv-cache`` mech move.

Cut ``Scheduler.build_kv_cache`` (@staticmethod after the prep commit)
and append to ``mem_cache/kv_cache_builder.py``. Drop ``@staticmethod``
decorator; dedent body to module level. Body bytes byte-equivalent.

Sole caller in ``Scheduler.__init__`` updated from
``Scheduler.build_kv_cache(...)`` → ``kv_cache_builder.build_kv_cache(...)``
(pure prefix replacement).

Refactor chain ID: extract-build-kv-cache-move
